refactor(tests): Extract `FeedLoaderStub` into shared scope to remove duplication
